WIP: scratch patch for migration
This is a scratchpad patch, which will probably be folded.

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/statistics/services/direct/multilayer/MeterDirectStatisticsServiceTest.java
@@ -36,7 +36,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yangtools.yang.common.Uint32;
 
 public class MeterDirectStatisticsServiceTest extends AbstractDirectStatisticsServiceTest {
-    static final Long METER_NO = 1L;
+    static final Uint32 METER_NO = Uint32.valueOf(1);
     private MeterDirectStatisticsService service;
 
     @Override
